requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#
click==6.7                # via pip-tools
first==2.0.1              # via pip-tools
icalendar==3.11.4
invoke==0.12.2
jinja2==2.9.6
markdown==2.6.8
MarkupSafe==1.0           # via jinja2
nose==1.3.7
pip-tools==1.9.0
python-dateutil==2.6.0    # via icalendar
pytz==2017.2              # via icalendar
pyyaml==3.12
requests==2.14.2
six==1.10.0               # via pip-tools, python-dateutil
